backend/requirements.txt
Django==5.1.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.9.0
psycopg2-binary==2.9.10
pdfplumber==0.11.5
